examples/530-voice-agent-multi-provider-proxy-python/requirements.txt
deepgram-sdk==6.1.1
fastapi==0.135.3
starlette==1.0.0
uvicorn[standard]==0.34.0
httpx==0.28.1
python-dotenv==1.2.2
boto3==1.37.23
